requirements.txt
fastapi>=0.109
uvicorn[standard]>=0.22
pillow>=10.0
qrcode[pil]>=7.4
svgwrite>=1.4
requests>=2.31
pydantic>=2.6
pytest>=7.4
httpx>=0.25
